requirements.txt
fastapi==0.116.1
uvicorn==0.35.0
sqlalchemy==2.0.43
pydantic==2.11.7
pydantic-settings==2.10.1
python-multipart==0.0.20
streamlit==1.49.1
requests==2.32.4
numpy==2.3.2
pypdf==5.9.0
python-docx==1.2.0
faiss-cpu==1.12.0
pytest==8.4.1
